requirements.txt
asn1crypto==0.24.0
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
cryptography==2.3.1
edgegrid-python==1.1.1
idna==2.7
ndg-httpsclient==0.5.1
pyasn1==0.4.4
pycparser==2.19
pyOpenSSL==18.0.0
PyYAML==3.13
requests==2.20.1
six==1.11.0
urllib3==1.24.1
